MBM/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-crispy-forms==1.9.1
django-sendgrid-v5==0.8.1
django-summernote==0.8.11.6
future==0.18.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
newsapi-python==0.2.6
postgres==3.0.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pycryptodomex==3.9.8
PyJWT==1.7.1
pylint==2.5.3
python-http-client==3.2.7
pytz==2020.1
requests==2.24.0
sendgrid==6.4.4
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
stream-django==1.6.2
stream-python==3.5.2
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
